Move and rework tcpserver/io tests
